Rename src/test/ folder. Keep package name as lowercase and sperate pipeline and generic test function
