+++ genomix/genomix-pregelix/src/test/java/edu/uci/ics/genomix/pregelix/jobrun/BasicGraphCleanTestSuite.java
@@ -12,7 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.uci.ics.genomix.pregelix.JobRun;
+package edu.uci.ics.genomix.pregelix.jobrun;
 
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
