+++ genomix/genomix-pregelix/src/test/java/edu/uci/ics/genomix/pregelix/testhelper/TipAddVertex.java
@@ -1,4 +1,4 @@
-package edu.uci.ics.genomix.pregelix.TestHelper;
+package edu.uci.ics.genomix.pregelix.testhelper;
 
 import java.util.Arrays;
 import java.util.Iterator;
